refactor(test): neutralize api fixture helper naming
