+++ turbo/apps/api/src/signals/routes/__tests__/billing-usage-pack-credits.test.ts
@@ -12,7 +12,7 @@ import { clearMockNow, mockNow } from "../../../lib/time";
 import {
   deleteFeatureSwitchesForUser,
   updateFeatureSwitchesForUser,
-} from "./helpers/zero-feature-switches";
+} from "./helpers/feature-switches";
 import { createZeroRouteMocks } from "./helpers/zero-route-test";
 import { testUsageSettlementRoutes } from "../test-usage-settlement";
 import {
