+++ turbo/apps/api/src/signals/routes/__tests__/integrations-slack-status.test.ts
@@ -13,7 +13,7 @@ import {
   seedSlackOrgConnection$,
   seedSlackOrgInstallation$,
   type SlackIntegrationFixture,
-} from "./helpers/zero-integrations-slack";
+} from "./helpers/integrations-slack";
 import {
   createFixtureTracker,
   createZeroRouteMocks,
