+++ turbo/apps/api/src/signals/routes/__tests__/integrations-slack.test.ts
@@ -26,7 +26,7 @@ import {
   seedSlackOrgConnection$,
   seedSlackOrgInstallation$,
   type SlackIntegrationFixture,
-} from "./helpers/zero-integrations-slack";
+} from "./helpers/integrations-slack";
 import { integrationsSlackRoutes } from "../integrations-slack";
 
 const context = testContext();
