+++ turbo/apps/api/src/signals/routes/__tests__/teams-browser-connect.test.ts
@@ -16,7 +16,7 @@ import {
   teamsConnectFixture,
   teamsFixtureExternalId,
   type TeamsConnectFixture,
-} from "./helpers/zero-teams-connect";
+} from "./helpers/teams-connect";
 import { zeroTeamsConnectRoutes } from "../zero-teams-connect";
 
 const context = testContext();
