+++ turbo/apps/api/src/signals/routes/__tests__/teams-connect.test.ts
@@ -20,7 +20,7 @@ import {
   teamsFixtureExternalId,
   teamsMessageActivityForTest,
   type TeamsConnectFixture,
-} from "./helpers/zero-teams-connect";
+} from "./helpers/teams-connect";
 import { zeroTeamsConnectRoutes } from "../zero-teams-connect";
 
 const context = testContext();
