+++ turbo/apps/api/src/signals/routes/__tests__/teams-oauth.test.ts
@@ -20,7 +20,7 @@ import {
   setupTeamsConnectTestEnv,
   teamsConnectFixture,
   type TeamsConnectFixture,
-} from "./helpers/zero-teams-connect";
+} from "./helpers/teams-connect";
 import { zeroTeamsConnectRoutes } from "../zero-teams-connect";
 
 const context = testContext();
